feat(i18n): Add package into app
