+++ src/locale/en-US.json
@@ -1,0 +1,3 @@
+{
+  "hello": "hello world"
+}
